Temporary fix until SC.Query is able to parse [primative] negative numbers.
New records (which have not yet been persisted) are now assigned temporary IDs
that being at 1000000 and continue to increment as records are added. Previously
they had been assigned nagative numbers that continue to decrement as records
are added, but SC.Query isn't able to parse primative negative numbers in a
condition statement.

This is a hack that allows us to move to Sproutcore ToT until SC.Query is
fixed properly.